+++ app/src/main/java/com/techcourse/DispatcherServletInitializer.java
@@ -5,8 +5,8 @@ import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.interface21.DispatcherServlet;
 import com.interface21.web.WebApplicationInitializer;
+import com.interface21.webmvc.servlet.DispatcherServlet;
 
 /**
  * Base class for {@link WebApplicationInitializer}
