+++ mvc/src/main/java/com/interface21/webmvc/servlet/HandlerAdapter.java
@@ -1,9 +1,7 @@
-package com.interface21.webmvc.servlet.mvc.tobe.handlerAdapter;
+package com.interface21.webmvc.servlet;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-
-import com.interface21.webmvc.servlet.ModelAndView;
 
 public interface HandlerAdapter {
