+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/handlerMapping/HandlerKey.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe.handlerMapping;
+package com.interface21.webmvc.servlet.mvc.handlerMapping;
 
 import java.util.Objects;
