+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/handlerAdapter/HandlerAdapterRegistryTest.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe.handlerAdapter;
+package com.interface21.webmvc.servlet.mvc.handlerAdapter;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import com.interface21.webmvc.servlet.mvc.tobe.handlerMapping.HandlerExecution;
+import com.interface21.webmvc.servlet.mvc.handlerMapping.HandlerExecution;
 
 class HandlerAdapterRegistryTest {
